Added create sphere and animations of molecules

+++ program/tools/create_movie.cpp
@@ -17,7 +17,7 @@ int main(int args, char *argv[]) {
 	int cpus = atoi(argv[2]);
 	int timesteps = atoi(argv[3]);
 	
-	double *positions = new double[3*1000000];
+	float *positions = new double[3*1000000];
 	ofstream file ("movie.xyz", ios::out);
 	
 	ifstream **movie_files = new ifstream*[cpus];
@@ -34,7 +34,7 @@ int main(int args, char *argv[]) {
 		for(int cpu=0;cpu<cpus;cpu++) { 
 			int N;
 			movie_files[cpu]->read(reinterpret_cast<char*>(&N),sizeof(int));
-			movie_files[cpu]->read(reinterpret_cast<char*>(&positions[3*num_particles]),3*N*sizeof(double));
+			movie_files[cpu]->read(reinterpret_cast<char*>(&positions[3*num_particles]),3*N*sizeof(float));
 			num_particles += N;
 		}
